+++ setup.php
@@ -8,7 +8,7 @@ use GlpiPlugin\Brandpulse\Config as BrandpulseConfig;
 
 defined('GLPI_ROOT') or die('No direct access allowed');
 
-const PLUGIN_BRANDPULSE_VERSION = '0.1.32';
+const PLUGIN_BRANDPULSE_VERSION = '0.1.33';
 const PLUGIN_BRANDPULSE_MIN_GLPI = '11.0.0';
 const PLUGIN_BRANDPULSE_MAX_GLPI = '12.0.0';
 const PLUGIN_BRANDPULSE_MIN_PHP = '8.2.0';